backend/requirements.txt
fastapi
uvicorn[standard]
cryptography
Pillow
python-multipart
pydantic
typer
